Further cleanup app.py
