+++ cloud/connecting-algorithm/src/app.py
@@ -1,6 +1,4 @@
 import json
-import os
-import boto3
 import requests
 import datetime
 
@@ -8,15 +6,11 @@ import constants
 import database
 import match_logic
 
-# SET THIS TO YOUR LOCAL OPERATING SYSTEM
-LOCAL_OPERATING_SYSTEM = 'windows'  # set to 'mac', 'windows' or 'linux'
-
-
 def lambda_handler(event, context):
     matches = []
     
     try:
-        client = get_client()
+        client = database.get_client()
 
         # Get all users from database
         users = database.getAllUsers(client)
@@ -52,29 +46,3 @@ def lambda_handler(event, context):
             "message": matches,
         }),
     }
-
-def get_client():
-    isLocalEnvironment = os.environ['ENVIRONMENT'] == 'local'
-
-    endpoint_url = "http://docker.for.mac.localhost:8002/"
-    if (isLocalEnvironment):
-        if LOCAL_OPERATING_SYSTEM == 'linux':
-            # linux
-            endpoint_url = "http://127.0.0.1:8002/"
-        elif LOCAL_OPERATING_SYSTEM == 'mac':
-            # OS X
-            endpoint_url = "http://docker.for.mac.localhost:8002/"
-        elif LOCAL_OPERATING_SYSTEM == 'windows':
-            # Windows
-            endpoint_url = "http://docker.for.windows.localhost:8002/"
-
-    # Create dynamo client based on environment
-    client = boto3.client(
-        'dynamodb',
-        region_name='us-east-1' if isLocalEnvironment else 'ap-southeast-2',
-        endpoint_url=endpoint_url if isLocalEnvironment else None,
-        aws_access_key_id='localKey' if isLocalEnvironment else None,
-        aws_secret_access_key='localSecret' if isLocalEnvironment else None,
-    )
-
-    return client
